commonize logic
